+++ CHANGELOG.md
@@ -9,16 +9,21 @@ Full documentation for rocSOLVER is available at [rocsolver.readthedocs.io](http
 - Reductions of generalized symmetric- and hermitian-definite eigenproblems:
     - SYGS2, SYGST (with batched and strided\_batched versions)
     - HEGS2, HEGST (with batched and strided\_batched versions)
+- Symmetric and Hermitian matrix Eigensolvers
+    - SYEV (with batched and strided\_batched versions)
+    - HEEV (with batched and strided\_batched versions)
 
 ### Optimizations
 
 ### Changed
+- Sorting method in STERF as original quick-sort was failing for large sizes.
 
 ### Removed
 - Removed hcc compiler support
 
 ### Fixed
 - Fixed GELS overwriting B even when info != 0
+- Error when calling STEQR with n=1 from batched routines
 
 
 ## [rocSOLVER 3.11.0 for ROCm 4.1.0]